Full calendar item update support
CalendarItem#update_item! did not support the full range of options
available via the EwsBuilder#build_xml! method as :text was the only
reserved attribute preserved and a nested structure was not supported at
all.

This commit makes :sub_elements and :xmlns_attribute reserved too and
ensures that any nested structures are also escaped as necessary.