Update calendar item correctly
Update calendar item was not generating required_attendees structure
correctly, since it was using `build_xml!`, which requires annoying
`sub_elements` in a lot of places. But the ews_builder actually have
`calendar_item!` method already (which is used in the `create_item`, but
not `update_item` - not entirely sure why).

This is a minimal change to directly use `calendar_item!` instead of
unintuitive `build_xml!` for updates too.