Add support for configuring a HTTP proxy for EWS requests
Currently, you have to subclass Viewpoint::EWS::Connection to be able to
gain access to the HTTP client so that a proxy URL and proxy authentication
can be configured.  This commit adds options that can be passed to
EWSClient.new to do this directly.
It also supports setting the proxy credentials later using the
`#set_proxy_auth` call on the connection.