Allow swapping out `HTTPClient` class
Previous changes to this library allowed swapping out the
`Viewpoint::EWS::Connection` class, but so far the use of the HTTPClient
was directly hard-coded into the default `Connection` class.

This change will make it easier to swap in alternative implementations
of HTTPClient without having to swap out the `Connection` class.