Handle special folder element called 'item'.
The problem is that the parser tries to instantiate a class
in Viewpoint::EWS::Types namespace that has the same name as
the element. In case of "item", this is not possible because
Item is a module, not a class. But we can redirect it to a
class called GenericItem.